+++ Mall/src/main/java/com/example/demo/model/PmsProductFullReduction.java
@@ -2,8 +2,8 @@ package com.example.demo.model;
 
 public class PmsProductFullReduction {
 	private Double fullPrice;
-	private Integer id;
-	private Integer productId;
+	private Long id;
+	private Long productId;
 	private Double reducePrice;
 	
 }
